# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Genesis Mesh Trust Gateway",
    "version": "0.56.1",
    "description": "Operator-controlled Genesis Mesh trust verification. HTTP 200 is not a trust grant; inspect trusted in each decision."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Signature": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "key_id",
          "sig"
        ],
        "properties": {
          "key_id": {
            "type": "string"
          },
          "sig": {
            "type": "string",
            "description": "Base64 detached Ed25519 signature"
          }
        }
      },
      "JoinCertificate": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "cert_id",
          "node_public_key",
          "network_name",
          "issued_at",
          "expires_at",
          "issued_by"
        ],
        "properties": {
          "cert_id": {
            "type": "string"
          },
          "node_public_key": {
            "type": "string"
          },
          "network_name": {
            "type": "string"
          },
          "issued_by": {
            "type": "string"
          },
          "issued_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "signatures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Signature"
            }
          }
        }
      },
      "Decision": {
        "type": "object",
        "required": [
          "trusted",
          "reasons"
        ],
        "properties": {
          "trusted": {
            "type": "boolean"
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/networks": {
      "get": {
        "summary": "Authorized network data",
        "operationId": "v1_networks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Bearer credential required"
          },
          "403": {
            "description": "Scope denied"
          },
          "429": {
            "description": "Quota exhausted"
          },
          "503": {
            "description": "Unavailable"
          }
        }
      }
    },
...
    "/verify": {
      "post": {
        "summary": "Verify a join certificate",
        "operationId": "verify",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Decision"
                }
              }
            }
          },
          "401": {
            "description": "Bearer credential required"
          },
          "403": {
            "description": "Scope denied"
          },
          "429": {
            "description": "Quota exhausted"
          },
          "503": {
            "description": "Unavailable"
          },
          "400": {
            "description": "Invalid request"
          },
          "408": {
            "description": "Request timeout"
          },
          "413": {
            "description": "Body too large"
          },
          "422": {
            "description": "Invalid JSON schema"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "certificate"
                ],
                "properties": {
                  "certificate": {
                    "$ref": "#/components/schemas/JoinCertificate"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/verify/batch": {
      "post": {
        "summary": "Verify a bounded batch",
        "operationId": "verify_batch",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Decision"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Bearer credential required"
          },
          "403": {
            "description": "Scope denied"
          },
          "429": {
            "description": "Quota exhausted"
          },
          "503": {
            "description": "Unavailable"
          },
          "400": {
            "description": "Invalid request"
          },
          "408": {
            "description": "Request timeout"
          },
          "413": {
            "description": "Body too large"
          },
          "422": {
            "description": "Invalid JSON schema"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "certificates"
                ],
                "properties": {
                  "certificates": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                      "$ref": "#/components/schemas/JoinCertificate"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}